requirements.txt
altair==4.1.0
astor==0.8.1
base58==2.1.0
blinker==1.4
cachetools==4.2.2
click==7.1.2
gitdb==4.0.7
GitPython==3.1.18
numpy==1.19.5
protobuf==3.17.3
pyarrow==5.0.0
pydeck==0.6.2
scipy==1.7.0
joblib==1.0.1
threadpoolctl==2.2.0
Cython==0.29.24
matplotlib==3.4.2
matplotlib-inline==0.1.2
pandas==1.3.0
seaborn==0.11.1
smmap==4.0.0
streamlit==0.85.1
toml==0.10.2
toolz==0.11.1
tzlocal==2.1
validators==0.18.2
Pillow==8.3.1